+++ lib/ipados_menu_bar.dart
@@ -9,6 +9,7 @@ part 'widgets/default_menus/edit_menu.dart';
 part 'widgets/default_menus/format_menu.dart';
 part 'widgets/default_menus/view_menu.dart';
 part 'widgets/default_menus/window_menu.dart';
+part 'widgets/default_menus/app_menu.dart';
 
 part 'widgets/icon/item_with_icon.dart';
 part 'widgets/icon/submenu_with_icon.dart';