+++ lib/widgets/default_menus/file_menu.dart
@@ -1,28 +1,12 @@
 part of '../../ipados_menu_bar.dart';
 
-/// Custom [PlatformMenu] using a menuId to replace the native "file" menu with
-/// the items from this class. Making it easier and faster to link callbacks
-/// from the app.
+/// Custom [PlatformMenu] using a menuId to access the native "file" menu adding
+/// thea additional items from this class. In case you need to listen when the
+/// app is closed, checking out [WidgetsBindingObserver] is highly recommended.
 class IPadFileMenu extends IPadMenu {
   @override
   String get menuId => 'file';
 
-  IPadFileMenu({
-    VoidCallback? onCloseWindow,
-    Intent? onCloseWindowIntent,
-    List<PlatformMenuItem>? additionalItems,
-  }) : super(
-         label: 'File',
-         menus: [
-           ...?additionalItems,
-           /* Should be logic only, don't render this item in the future */
-           PlatformMenuItemWithIcon(
-             icon: CupertinoIcons.xmark,
-             label: 'Close Window',
-             onSelectedIntent: onCloseWindowIntent,
-             onSelected: onCloseWindow,
-             shortcut: SingleActivator(LogicalKeyboardKey.keyW, meta: true),
-           ),
-         ],
-       );
+  IPadFileMenu({List<PlatformMenuItem>? additionalItems})
+      : super(label: 'File', menus: [...?additionalItems]);
 }